reduce class stack
